BRS-458: updated text on beta banner and componentized banner
